% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingfunctions.R
\name{InteRD.predict.prop}
\alias{InteRD.predict.prop}
\title{Integrated and robust deconvolution}
\usage{
InteRD.predict.prop(InteRD.output)
}
\arguments{
\item{InteRD.output}{An object from InteRD1 or InteRD2.}
}
\value{
Estimated cell type proportions via InteRD.
}
\description{
This function provides a robust devonvolution framework to integrate information from scRNA-seq references, marker genes, and prior biological knowledge.
}
