% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRD.R
\name{InteRD1}
\alias{InteRD1}
\title{Reference ensemble}
\usage{
InteRD1(bulk.data,list_marker,cell_type_unique,comb_used,lambda_option)
}
\arguments{
\item{bulk.data}{ExpressionSet object for a target bulk data.}

\item{list_marker}{A list of pre-specified marker genes corresponding to each cell type.}

\item{cell_type_unique}{A list of cell types. It should match the order in list.marker.}

\item{comb_used}{A list of pre-estimated cell type proportions based on different references.}

\item{lambda_option}{a sequence of values for the tunning parameter.}
}
\value{
A list containing estimated cell type proportions corresponding to each tuning value, named est; and a sequence of goodness-of-fit values corresponding to  each tuning value, named metrics. The smaller the better; and a list of weights corresponding to each tuning value, named weights_list.
}
\description{
This function provides a reference-based deconvolution by resembling all estimated cell-type proportions based on each reference set.
}
