% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingfunctions.R
\name{pop.ct.prop.scRNA}
\alias{pop.ct.prop.scRNA}
\title{Calculate population-level cell type proportions from single-cell data.}
\usage{
pop.ct.prop.scRNA(scRNA,cluster="cluster",sample="sample",cell_type_unique)
}
\arguments{
\item{scRNA}{ExpressionSet object for single-cell data.}

\item{cluster}{Character string specifying the variable name for cell types. The default is "cluster".}

\item{sample}{Character string specifying the variable name for subject/samples. The default is "sample".}

\item{cell_type_unique}{A vector of cell types. It should match the order in list.marker.}
}
\value{
Population-level cell type proportions (pop.ct.prop) and corresponding standard deviations (pop.ct.sd).
}
\description{
Calculate population-level cell type proportions from single-cell data.
}
