% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRD.R
\name{RF}
\alias{RF}
\title{A reference-free deconvolution approach}
\usage{
RF(bulk.data,list_marker,cell_type_unique)
}
\arguments{
\item{bulk.data}{ExpressionSet object for a target bulk data.}

\item{list_marker}{A list of pre-specified marker genes corresponding to each cell type.}

\item{cell_type_unique}{A list of cell types. It should match the order in list.marker.}
}
\value{
Estimated cell type proportions, named est; and a goodness-of-fit value, named metrics. The smaller the better.
}
\description{
This function provides a reference-free devonvolution framework.
}
