% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingfunctions.R
\name{InteRD.predict.prop}
\alias{InteRD.predict.prop}
\title{Extract the estimated proportions from InteRD}
\usage{
InteRD.predict.prop(InteRD.output)
}
\arguments{
\item{InteRD.output}{An object from InteRD1 or InteRD2.}
}
\value{
Estimated cell type proportions from InteRD.
}
\description{
This function extract estimated cell type proportions via InteRD1 and InteRD2.
}
\examples{
##read data
library(InteRD)
readRDSFromWeb<-function(ref) {readRDS(gzcon(url(ref)))}
urlremote<-"https://github.com/chencxxy28/Data/raw/main/data_InteRD/"
pseudo.seger<-readRDSFromWeb(paste0(urlremote,"pseudo.seger.rds"))
comb<-readRDSFromWeb(paste0(urlremote,"comb_seger.rds"))
list_marker<-readRDSFromWeb(paste0(urlremote,"list_markerbaron20.rds"))
lambda_option<-0
cell_type_unique<-c("alpha","beta","delta","gamma")
InteRD1.output<-InteRD1(bulk.data =pseudo.seger$pseudo_eset,list_marker,
cell_type_unique,comb_used=comb,lambda_option)
InteRD1<-InteRD.predict.prop(InteRD.output=InteRD1.output)

}
