% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingfunctions.R
\name{InteRD.predict.prop}
\alias{InteRD.predict.prop}
\title{Extract the estimated proportions from InteRD}
\usage{
InteRD.predict.prop(InteRD.output)
}
\arguments{
\item{InteRD.output}{An object from InteRD1 or InteRD2.}
}
\value{
Estimated cell type proportions from InteRD.
}
\description{
This function extract estimated cell type proportions via InteRD1 and InteRD2.
}
